feat: implement layered Ethernet network diagnostics (L1-L5 checkers) .
